added the creation of a rotation figure

+++ AffineTransforms_3D/AffineTransforms.cs
@@ -46,7 +46,7 @@ namespace AffineTransforms_3D
       
 
 
-        static public Transformator RotateTransform3D(Point3D center, int angle, double x=0, double y=0, double z=0)
+        static public Transformator RotateTransform3D(Point3D center, double angle, double x=0, double y=0, double z=0)
         {
             var rotator = new RotateTransform3D(new AxisAngleRotation3D(new Vector3D(x, y, z), angle),center);
             return new StandardMatrixTransformator(rotator);
